requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==3.0.3
idna==3.4
Jinja2==3.1.4
MarkupSafe==2.1.5
playwright==1.45.0
pyaml==24.4.0
pyee==11.1.0
PyYAML==6.0.1
requests==2.31.0
typing_extensions==4.12.2
urllib3==2.0.4
